Fix detection of fail2ban-server process
We discovered a corner case where icinga on an old system (debian jessie) is started with COLUMNS=100 in its environment. This cuts off the output of the `ps -aux` command used for testing if fail2ban-server is running, rendering the detection useless. The server process is always reported as not running, even if it is running. Simply setting COLUMNS to an empty value prevents this bug, which may be triggered otherwise, too.